Refactor document validation
